apply feedback from bitterpanda63
